requirements.txt
bs4
pandas
newspaper3k
nltk
lxml[html_clean]
